Refactor Sleep pure logic helpers
