+++ android/app/src/main/java/com/noop/ui/SleepScreen.kt
@@ -92,15 +92,12 @@ import com.noop.data.DismissedSleep
 import com.noop.data.SleepSession
 import com.noop.data.WhoopRepository
 import kotlinx.coroutines.launch
-import org.json.JSONArray
-import org.json.JSONObject
 import java.time.LocalDate
 import java.time.format.DateTimeFormatter
 import java.text.SimpleDateFormat
 import java.util.Calendar
 import java.util.Date
 import java.util.Locale
-import java.util.TimeZone
 import kotlin.math.abs
 import kotlin.math.max
 import kotlin.math.roundToInt
@@ -2904,680 +2901,7 @@ private fun SleepEmptyState() {
 
 // MARK: - Model + derivation (faithful to SleepView.swift)
 
-/** Stage minutes for a single night (mirrors the macOS Stages struct). */
-internal data class Stages(
-    val awake: Double,
-    val light: Double,
-    val deep: Double,
-    val rem: Double,
-) {
-    /** Total time in bed (includes awake). */
-    val total: Double get() = awake + light + deep + rem
-
-    /** Asleep time = total minus awake. */
-    val asleep: Double get() = light + deep + rem
-}
-
-/** (latest, typical mean, full history) per metric — mirrors the macOS Metric tuple. */
-internal data class Metric(
-    val latest: Double?,
-    val typical: Double?,
-    val series: List<Double>,
-)
-
-/** Export-verbatim per-day sleep figures (metricSeries keys mirroring macOS WhoopImporter). */
-internal data class ImportedSleepSeries(
-    val performance: Map<String, Double> = emptyMap(), // sleep_performance, 0–100
-    val consistency: Map<String, Double> = emptyMap(), // sleep_consistency, 0–100
-    val needMin: Map<String, Double> = emptyMap(),     // sleep_need_min, minutes
-    val debtMin: Map<String, Double> = emptyMap(),     // sleep_debt_min, minutes
-)
-
-/** Everything the screen renders, derived once per data change. */
-internal data class SleepModel(
-    val stages: Stages,
-    val clockLabel: String,
-    val efficiencyText: String,
-    val performance: Metric,
-    val efficiency: Metric,
-    val consistency: Metric,
-    val hoursVsNeeded: Metric,
-    val restorative: Metric,
-    val respiratory: Metric,
-    val sleepDebt: Metric,
-    val typicalTotalMin: Double?,
-    val typicalDeepMin: Double?,
-    val typicalRemMin: Double?,
-    val typicalLightMin: Double?,
-    val trendHours: List<Double>,
-    val trendNeedHours: List<Double>,
-    val trendDebtHours: List<Double>,
-    val trendDates: List<String>,
-    /** Persisted per-epoch segments as ordered (stage, minutes) weights — the REAL
-     *  hypnogram (on-device APPROXIMATE staging) — or null → synthesized fallback. */
-    val realSegments: List<Pair<String, Float>>?,
-    /** Rolling 14-night sleep-debt ledger: Σ(slept − personal need) across the recent
-     *  fortnight, with the per-night deltas behind it. Computed once per data change. (#242) */
-    val sleepDebtLedger: SleepDebtLedger,
-)
-
-/** The night the ◀/▶ chevrons selected: its MAIN session, the day-metric key it resolves to, its
- *  persisted per-epoch weights (or null), the "EEE d MMM · HH:mm–HH:mm" clock, and the day's other
- *  blocks (naps / split-sleep) for the naps card. (#160, #518) */
-internal data class HeroNight(
-    val session: SleepSession,
-    val dayKey: String,
-    val realSegments: List<Pair<String, Float>>?,
-    val clockLabel: String,
-    val napBlocks: List<SleepSession> = emptyList(),
-    // The bridged main-night GROUP (#561): summed stage minutes + the full-night segments, when the night
-    // is more than one fragment. `session` above stays the single WINNING block (the edit anchor); these
-    // let buildSleepModel render the WHOLE night instead of one fragment (#555). Null for a single-block day.
-    val groupStages: StageMins? = null,
-    val groupSegments: List<PersistedSegment>? = null,
-    // Per-epoch MOTION for the main-night GROUP (#407), laid fragment-by-fragment in the SAME order the
-    // group's stage segments are laid. Empty when no group fragment has a persisted motionJSON (older rows)
-    // → the hero shows an honest empty state instead of a fabricated zero trace. Read off the already-
-    // resolved group, NOT a re-resolution of the night.
-    val groupMotion: List<Double> = emptyList(),
-    // Time-in-bed for the whole main-night GROUP (#561): Σ(endTs − effectiveStartTs) across the
-    // hero fragments, in minutes. The hero subtitle previously derived in-bed from `session` alone
-    // (the single WINNING fragment), so a fragmented night's "Xh in bed" undershot the stage total
-    // it sat next to. Summing fragment windows (NOT wall-clock first-onset→last-wake) excludes the
-    // inter-fragment awake gaps, mirroring sumGroupStages/AnalyticsEngine, so asleep ≤ in-bed and
-    // the efficiency shown beside it stays coherent. Null for a single-block day → the hero keeps
-    // its session-window / stage-total fallbacks.
-    val groupInBedMin: Double? = null,
-    // The whole bridged night's clock WINDOW (#345): the displayed bedtime (first non-stub fragment's
-    // onset, #736) to the group's latest wake — the same pair `clockLabel` above is built from, carried
-    // as timestamps so the Asleep/Woke row + the hypnogram axis can use them. On a split night `session`
-    // (the single WINNING fragment, kept as the edit anchor) can end mid-night, so reading ITS endTs
-    // made the WOKE time + the axis contradict the header pill and the group hypnogram. iOS needs no
-    // analogue: its merged Night synthesizes a group-spanning session (mergeDay's `synth`), so its
-    // window row and axis were already whole-night. Null only via the default → session fallback.
-    val heroOnsetTs: Long? = null,
-    val heroWakeTs: Long? = null,
-)
-
-/** What the hero card draws for the selected night — null means no usable stage data
- *  (renders the honest "No stage data recorded for this night." fallback). (#160) */
-internal data class HeroDisplay(
-    val stages: Stages,
-    val realSegments: List<Pair<String, Float>>?,
-    val efficiencyText: String,
-)
-
-/**
- * Pick the night for the DAY [offset] stops back from the most recent (0 = latest). [navDays]
- * is grouped-by-calendar-day, newest first, so the chevrons step by DAY not by flat session
- * index — a WHOOP 4.0 user with a single detected night has exactly one stop (both arrows
- * correctly disabled) instead of arrows that move within naps/split blocks of one night and
- * appear stuck (#57/#59). Mirrors iOS SleepView.decodedNight(at:)/navDays.
- *
- * The day's REPRESENTATIVE session is its MAIN sleep block — the LONGEST block, preferring an
- * OVERNIGHT-anchored onset (#518). A day can hold an overnight AND an afternoon nap (both end on
- * the same calendar day, so both bucket here); the OLD `maxByOrNull { endTs }` picked the
- * latest-ending block, which is the afternoon nap — so the overnight vanished from the Sleep tab.
- * Picking the longest overnight block fixes it; the other blocks are carried as `napBlocks` for
- * the naps card. The day key tries UTC then local-tz attribution of the MAIN block's wake — imported
- * DailyMetric.day is local-tz while dayString is UTC, so a near-midnight-UTC wake needs the second
- * key; both derive from THIS night's endTs, never another night. (#160, #518)
- */
-internal fun selectNight(
-    navDays: List<List<SleepSession>>,
-    days: List<DailyMetric>,
-    offset: Int,
-    // The LEARNED habitual midsleep the engine threaded into the daily total, so the hero, the naps split,
-    // and the edit target pick the SAME block the analytics rollup did — for a shift/late sleeper too. null
-    // = cold-start band. (#547)
-    habitualMidsleepSec: Long? = null,
-    // Per-epoch MOTION keyed by detected startTs (#407). The group's fragments' series are concatenated in
-    // group order onto HeroNight.groupMotion. Empty/absent → honest empty state. Default empty so existing
-    // callers/tests compile unchanged.
-    motionByStart: Map<Long, List<Double>> = emptyMap(),
-): HeroNight? {
-    if (navDays.isEmpty()) return null
-    val dayIdx = offset.coerceIn(0, navDays.size - 1)
-    val blocks = navDays[dayIdx]
-    val session = mainSleepBlock(blocks, habitualMidsleepSec) ?: return null
-    // The day's MAIN sleep is the bridged main-night GROUP (#561): a briefly-interrupted / biphasic night's
-    // sibling fragments belong to the night, NOT the naps card — only blocks OUTSIDE the group are naps.
-    // `session` stays the single WINNING block (the durable-edit anchor at SleepTimeEditor), but the group
-    // drives the naps split, the hero's summed stage minutes, and the full-night hypnogram, so the tab
-    // matches AnalyticsEngine.analyzeDay instead of rendering phantom naps (#555). A single-block day is
-    // byte-identical to the prior behaviour. (#518/#555/#561)
-    val group = mainSleepGroup(blocks, habitualMidsleepSec)
-    val groupStarts = group.map { it.startTs }.toHashSet()
-    val napBlocks = blocks.filter { it.startTs !in groupStarts }
-        .sortedBy { it.effectiveStartTs }
-    // Drop a spurious leading pre-sleep awake stub from the hero's RECONSTRUCTION so the hypnogram and the
-    // summed minutes start where the displayed bedtime (the main block's onset) does (#736). A night can
-    // record a brief, all-awake pre-onset block (e.g. lying in bed before sleep); the gap-bridge folds it
-    // into the group, so the chart drew sleep beginning before the labelled "Asleep" time. We only drop a
-    // BRIEF, essentially-sleepless leading fragment that also sits before the main block, so a genuine first
-    // sleep fragment of an interrupted/biphasic night is never lost. The stub still rides in `groupStarts`
-    // above, so it is never mislabelled as a nap. `session` (the edit anchor) is already the main block, so
-    // the bedtime label and the pencil were aligned — this aligns the chart to that same bedtime. (#736/#555)
-    val onsetTsForHero = session.effectiveStartTs
-    // #259: reference size for the "minor relative to the main block" stub test = the largest asleep span in
-    // the group (≈ the main block). A genuine biphasic first sleep is comparable to it and is kept; only a
-    // small stray lead carrying a few minutes of sleep is dropped, so the onset no longer jumps hours early.
-    val groupRefAsleepMin = group.maxOfOrNull { frag ->
-        decodedAsleepMinutes(frag.stagesJSON, frag.effectiveStartTs)
-    } ?: 0.0
-    val heroGroup = group.dropWhile {
-        it.effectiveStartTs < onsetTsForHero && isPreOnsetAwakeStub(it, groupRefAsleepMin)
-    }
-    val utcKey = AnalyticsEngine.dayString(session.endTs)
-    val localKey = localDayString(session.endTs)
-    val dayKey = listOf(utcKey, localKey).firstOrNull { key ->
-        days.any { it.day == key && (it.deepMin ?: 0.0) + (it.remMin ?: 0.0) + (it.lightMin ?: 0.0) > 0.0 }
-    } ?: utcKey
-    // Lay every fragment's persisted segments end-to-end so a biphasic night draws as one continuous
-    // hypnogram, and SUM their stage minutes for the hero. Built from `heroGroup` (the group minus a leading
-    // spurious stub, #736) so the chart and minutes start at the displayed bedtime. Null for a single-block
-    // hero → prior behaviour.
-    // #259: clamp each fragment's timeline to its effective onset too (matching the stage-minute clamp
-    // and the iOS decodeSegments clamp), so the hypnogram starts where the edited bedtime does instead of
-    // drawing pre-onset bars that contradict the clamped asleep total. No-op for non-edited nights.
-    val groupSegmentsRaw = if (heroGroup.size > 1) {
-        heroGroup.flatMap { parsePersistedSegments(SleepStageTotals.clampStagesToOnset(it.stagesJSON, it.effectiveStartTs)).orEmpty() }
-            .sortedBy { it.start }
-            .takeIf { it.size >= 2 }
-    } else null
-    // #364: draw each inter-fragment wake seam as an explicit wake segment in the full-night
-    // hypnogram, so the merged night has no silent hole where the user was up — matching what the
-    // Health Connect export now writes. TIMELINE only: the seam is deliberately NOT added to the
-    // stage MINUTES (sumGroupStages) or groupInBedMin below — those keep the fragment-only
-    // accounting (asleep ≤ in-bed, see the #561 comment), and iOS mergeDay applies the same split.
-    val groupSegments = groupSegmentsRaw?.let { segs ->
-        val seams = heroGroup.zipWithNext().mapNotNull { (prev, next) ->
-            if (next.effectiveStartTs > prev.endTs)
-                PersistedSegment(prev.endTs, next.effectiveStartTs, "wake") else null
-        }
-        (segs + seams).sortedBy { it.start }
-    }
-    val groupStages = if (heroGroup.size > 1) sumGroupStages(heroGroup) else null
-    val segments = (groupSegmentsRaw ?: parsePersistedSegments(SleepStageTotals.clampStagesToOnset(session.stagesJSON, session.effectiveStartTs)))
-        ?.map { seg -> seg.stage to ((seg.end - seg.start) / 60f) }
-    // #407: lay the GROUP's per-epoch motion fragment-by-fragment in `heroGroup` order (the same order
-    // `groupSegments` lays the stage timeline), reading the already-chosen group's stored series. The
-    // detected key (`startTs`) is the motion store's key. A fragment with no series contributes nothing; if
-    // NO fragment has one, `groupMotion` is empty → honest empty state.
-    val groupMotion = heroGroup.flatMap { motionByStart[it.startTs].orEmpty() }
-    // #736 parity: the displayed bedtime must match where the hypnogram starts. The chart is built from
-    // heroGroup (first non-stub fragment onward), so label from THAT fragment's onset (mirrors Swift
-    // nightOnsetTs / synth.startTs), closed by the group's latest wake. `session` stays the edit anchor only.
-    val heroOnsetTs = heroGroup.firstOrNull()?.effectiveStartTs ?: session.effectiveStartTs
-    val heroWakeTs = heroGroup.maxOfOrNull { it.endTs } ?: session.endTs
-    // #561: whole-group time-in-bed (minutes) — fragment windows summed, gaps excluded — so the hero
-    // subtitle matches the multi-fragment stage total it is shown with. Single-block days stay null.
-    val groupInBedMin = if (heroGroup.size > 1) {
-        heroGroup.sumOf { (it.endTs - it.effectiveStartTs).coerceAtLeast(0L) } / 60.0
-    } else null
-    return HeroNight(session, dayKey, segments, clockLabelFor(heroOnsetTs, heroWakeTs), napBlocks, groupStages,
-        groupSegments, groupMotion, groupInBedMin, heroOnsetTs, heroWakeTs)
-}
-
-/**
- * The day's MAIN sleep block — the night people mean by "last night" — resolved by the SINGLE shared
- * selector ([SleepStageTotals.mainNightIndex]) the analytics rollup uses: the LEARNED-TIMING score
- * (asleep span + alignment bonus on each block's EFFECTIVE onset) rather than a re-derived overnight
- * gate, so the hero, the edit affordance, the analytics total, and the Sleep tab ALL resolve to the
- * identical block (the whole point of #525/#547). Scores on each block's EFFECTIVE onset (what the user
- * sees) and returns the owning session. This is the BARE single-block pick (the durable-edit anchor): the
- * HERO display and the nap split use [mainSleepGroup], which bridges the winner's adjacent fragments into
- * ONE night (#561) so a biphasic night isn't shown as phantom naps (#555). [habitualMidsleepSec] is the
- * SAME learned value the engine threads into the
- * persisted totals (loaded via `vm.repo.habitualMidsleepSec`), so a shift/late sleeper's hero and analytics
- * total resolve to the identical block; null keeps the cold-start overnight-band bonus, which matches a
- * cold-start engine run. Mirrors iOS SleepView.mainNightSession. (#518/#547)
- */
-internal fun mainSleepBlock(blocks: List<SleepSession>, habitualMidsleepSec: Long? = null): SleepSession? {
-    if (blocks.isEmpty()) return null
-    val idx = SleepStageTotals.mainNightIndex(
-        blocks.map { SleepStageTotals.NightBlock(it.effectiveStartTs, it.endTs) },
-        uiTzOffsetSec(),
-        habitualMidsleepSec,
-    ) ?: return null
-    return blocks[idx]
-}
-
-/**
- * The day's MAIN-night GROUP — the winning block PLUS any adjacent fragments bridged into it (a wake gap
- * shorter than [SleepStageTotals.gapBridgeMaxMin]), so a briefly-interrupted / biphasic night reads as ONE
- * continuous sleep exactly the way AnalyticsEngine.analyzeDay rolls it up for the daily total (#561). The
- * hero aggregates this whole group and ONLY blocks outside it are naps; without it the tab used the
- * un-bridged single-block pick and rendered the bridged siblings as phantom naps (#555). A night with no
- * bridgeable gap collapses to the single block [mainSleepBlock] picks. Returns ascending by effective
- * onset. Mirrors iOS SleepView.mainNightGroup. (#561/#555)
- */
-internal fun mainSleepGroup(blocks: List<SleepSession>, habitualMidsleepSec: Long? = null): List<SleepSession> {
-    val idx = SleepStageTotals.mainNightGroupIndices(
-        blocks.map { SleepStageTotals.NightBlock(it.effectiveStartTs, it.endTs) },
-        uiTzOffsetSec(),
-        habitualMidsleepSec,
-    ) ?: return emptyList()
-    return idx.map { blocks[it] }.sortedBy { it.effectiveStartTs }
-}
-
-/**
- * The day's main-night bridged SPAN (onset -> wake), the same window [mainSleepGroup] bridges into one
- * continuous night. The ONE canonical bed/wake read every glance screen (Coupled, Today's HR band) should
- * show -- never a screen-local "freshest" or "longest single block" heuristic, which can silently disagree
- * with each other and with the Sleep tab hero on a night stored as more than one block (#294). null only
- * when `blocks` has nothing bridgeable. Mirrors iOS SleepView.mainNightSpan.
- */
-internal fun mainSleepSpan(blocks: List<SleepSession>, habitualMidsleepSec: Long? = null): Pair<Long, Long>? {
-    val group = mainSleepGroup(blocks, habitualMidsleepSec)
-    val first = group.firstOrNull() ?: return null
-    val last = group.lastOrNull() ?: return null
-    return first.effectiveStartTs to last.endTs
-}
-
-/** Longest a leading block can be and still be treated as a spurious pre-sleep awake stub (lying in bed
- *  before sleep). Generous (a few hours) because the reporter's stub ran 21:41 → 00:27 — ~2h45m of pre-sleep
- *  awake — so a tight cap missed it (#736). The real guard against swallowing a genuine first sleep fragment
- *  is [PRE_ONSET_STUB_ASLEEP_MAX_MIN]: a stub must be essentially SLEEPLESS. Mirrors iOS
- *  SleepView.preOnsetStubMaxMin. (#736) */
-private const val PRE_ONSET_STUB_MAX_MIN = 240.0
-/** Most asleep minutes a fragment can carry and still count as a (sleepless) pre-onset awake stub. A real
- *  first sleep fragment of a biphasic night carries far more. Mirrors iOS SleepView.preOnsetStubAsleepMaxMin.
- *  (#736) */
-private const val PRE_ONSET_STUB_ASLEEP_MAX_MIN = 3.0
-/** A leading pre-onset fragment that carries SOME sleep is still spurious when it is minor RELATIVE to the
- *  night's main block: its asleep minutes are below this fraction of the largest fragment's. A genuine
- *  biphasic first sleep is comparable in size to the main block (well above this), so it is never dropped;
- *  only a small stray lead (e.g. a brief early doze hours before the real sleep) is. This extends the
- *  essentially-sleepless [PRE_ONSET_STUB_ASLEEP_MAX_MIN] rule (#736), which missed a lead carrying a few
- *  minutes more than 3. Mirrors iOS SleepView.preOnsetStubMinorFrac. (#259) */
-private const val PRE_ONSET_STUB_MINOR_FRAC = 0.15
-
-/** Absolute floor (ASLEEP minutes) under the #259 relative "minor lead" test: a leading fragment that carries
- *  at least this much real sleep is a genuine first sleep — a real sleep episode — and is NEVER a spurious
- *  pre-onset lead, however large the main block is. Without it a long main sleep inflates the 15% relative bar
- *  (a 6 h night → ~54 min) so a genuine ~34-min first sleep was swallowed and the shown bedtime jumped hours
- *  late, hiding the real onset the bridged night (and the Health write-back) already spans. 20 min ≈ the
- *  shortest standalone sleep episode; below it a handful of asleep minutes beside a long night is a stray lead.
- *  Mirrors iOS SleepView.preOnsetStubMinorAsleepFloorMin. (#259 / bridged-night headline) */
-internal const val PRE_ONSET_STUB_MINOR_ASLEEP_FLOOR_MIN = 20.0
-
-/**
- * Asleep minutes decoded from a stored [stagesJSON] in EITHER of the two formats that exist in the DB:
- * on-device COMPUTED nights store a SEGMENT ARRAY `[{start,end,stage}]`; imported nights store a dict of
- * MINUTES `{light,deep,rem,awake}`. [parseSessionStages] already reads both; this additionally threads the
- * fragment's [effectiveStartTs] through [SleepStageTotals.clampStagesToOnset] so a segment array is trimmed to
- * the effective onset (the #259 pre-onset trim) exactly as the hero's stage totals trim it — a no-op for a
- * minute dict and for a segment array that already starts at its onset. The displayed-onset stub test reads
- * asleep minutes through this seam so a computed night's segment array is never counted as 0 asleep minutes.
- * Internal so the onset golden pins the DECODE PATH itself. Mirrors iOS SleepView.decodedAsleepMinutes.
- */
-internal fun decodedAsleepMinutes(stagesJSON: String?, effectiveStartTs: Long): Double =
-    parseSessionStages(SleepStageTotals.clampStagesToOnset(stagesJSON, effectiveStartTs))
-        ?.let { it.light + it.deep + it.rem } ?: 0.0
-
-/** A fragment is a spurious pre-onset awake stub when it is within the lie-in cap (<= [PRE_ONSET_STUB_MAX_MIN])
- *  and EITHER carries essentially no sleep (asleep minutes <= [PRE_ONSET_STUB_ASLEEP_MAX_MIN]) OR is minor
- *  relative to the night's main block ([refAsleepMin], the group's largest asleep span): asleep minutes below
- *  [PRE_ONSET_STUB_MINOR_FRAC] of it AND below the absolute [PRE_ONSET_STUB_MINOR_ASLEEP_FLOOR_MIN] real-sleep-
- *  episode floor. Used only to skip such a stub when it leads the main-night group, so the hero's hypnogram and
- *  minutes start at the displayed bedtime (the main block's onset) rather than before it. [refAsleepMin]
- *  defaults to 0 (relative test off) so existing callers/tests are byte-identical. Mirrors iOS
- *  SleepView.isPreOnsetAwakeStub. (#736 / #259) */
-internal fun isPreOnsetAwakeStub(frag: SleepSession, refAsleepMin: Double = 0.0): Boolean {
-    val spanMin = (frag.endTs - frag.effectiveStartTs) / 60.0
-    if (spanMin > PRE_ONSET_STUB_MAX_MIN) return false
-    val asleepMin = decodedAsleepMinutes(frag.stagesJSON, frag.effectiveStartTs)
-    if (asleepMin <= PRE_ONSET_STUB_ASLEEP_MAX_MIN) return true
-    // #259 relative "minor lead" test, floored: a real sleep episode (>= the floor) is never a stray lead, so
-    // a long main block can't inflate the 15% bar past a genuine short first sleep. A genuine biphasic first
-    // sleep is comparable in size, so it stays and its onset stands.
-    return refAsleepMin > 0.0 &&
-        asleepMin < PRE_ONSET_STUB_MINOR_FRAC * refAsleepMin &&
-        asleepMin < PRE_ONSET_STUB_MINOR_ASLEEP_FLOOR_MIN
-}
-
-/** SUM the per-stage minutes across a bridged main-night group, so the hero's stage breakdown reflects the
- *  WHOLE night (#561) instead of one fragment (#555). The inter-fragment wake gap belongs to no fragment,
- *  so it is excluded exactly as AnalyticsEngine excludes it. Null if no fragment has parseable stages. */
-private fun sumGroupStages(group: List<SleepSession>): StageMins? {
-    var aw = 0.0; var li = 0.0; var dp = 0.0; var rm = 0.0; var any = false
-    for (frag in group) {
-        // #259: each fragment's stages trimmed to its effective onset before summing (see buildSleepModel).
-        val s = parseSessionStages(SleepStageTotals.clampStagesToOnset(frag.stagesJSON, frag.effectiveStartTs)) ?: continue
-        aw += s.awake; li += s.light; dp += s.deep; rm += s.rem; any = true
-    }
-    return if (any) StageMins(aw, li, dp, rm) else null
-}
-
-/** The device's current UTC offset (seconds east), evaluated per pick, fed to the selector's `offsetSec`
- *  so the timing test reads the user's clock via the SAME `offsetSec` math the engine uses
- *  ([SleepStageTotals.localSecOfDay]) instead of `Calendar.get(HOUR_OF_DAY)` — the duplicated, DST-fragile
- *  gate the audit flagged. Mirrors the engine's `TimeZone.getDefault().getOffset(...)`. (#547) */
-internal fun uiTzOffsetSec(): Long =
-    java.util.TimeZone.getDefault().getOffset(System.currentTimeMillis()) / 1000L
-
-/**
- * Resolve what the hero shows: the day-metric model when it resolved for the selected
- * night; else the session's own persisted segments (the day row can miss while the
- * segments exist); else null → the honest fallback. Never another night's data. (#160)
- */
-internal fun heroDisplay(model: SleepModel?, night: HeroNight?): HeroDisplay? {
-    if (model != null) return HeroDisplay(model.stages, model.realSegments, model.efficiencyText)
-    val segments = night?.realSegments ?: return null
-    val stages = stagesFromSegments(segments) ?: return null
-    val eff = night.session.efficiency
-        ?.let { e -> "${(if (e <= 1.0) e * 100.0 else e).roundToInt()}%" } ?: "—"
-    return HeroDisplay(stages, segments, eff)
-}
-
-/** Sum (stage, minutes) weights into per-stage totals; null when nothing is > 0. */
-internal fun stagesFromSegments(segments: List<Pair<String, Float>>): Stages? {
-    var awake = 0.0; var light = 0.0; var deep = 0.0; var rem = 0.0
-    for ((stage, minutes) in segments) {
-        val m = minutes.toDouble()
-        when (stage) {
-            "wake", "awake" -> awake += m
-            "light" -> light += m
-            "deep" -> deep += m
-            "rem" -> rem += m
-        }
-    }
-    val s = Stages(awake = awake, light = light, deep = deep, rem = rem)
-    return if (s.total > 0.0) s else null
-}
-
-internal data class StageMins(val awake: Double, val light: Double, val deep: Double, val rem: Double)
-
-/**
- * Extract stage minute counts from a session's stagesJSON, handling both formats:
- *  • Minute dict  {"awake":…,"light":…,"deep":…,"rem":…}  — imported nights (noopdb / WHOOP export)
- *  • Segment array [{start,end,stage}]                     — on-device computed nights
- * Returns null when the JSON is absent or unparseable, so callers fall back to DailyMetric columns.
- * SleepWindowReclip keeps the minute dict up to date after a wake-time edit, so stage counts
- * are correct immediately — no rescore needed for imported nights.
- */
-private fun parseSessionStages(stagesJSON: String?): StageMins? {
-    stagesJSON ?: return null
-    return runCatching {
-        val trimmed = stagesJSON.trim()
-        when {
-            trimmed.startsWith("{") -> {
-                val obj = JSONObject(trimmed)
-                val aw = obj.optDouble("awake", 0.0)
-                val li = obj.optDouble("light", 0.0)
-                val dp = obj.optDouble("deep", 0.0)
-                val rm = obj.optDouble("rem", 0.0)
-                if (aw + li + dp + rm > 0.0) StageMins(aw, li, dp, rm) else null
-            }
-            trimmed.startsWith("[") -> {
-                val arr = JSONArray(trimmed)
-                var aw = 0.0; var li = 0.0; var dp = 0.0; var rm = 0.0
-                for (i in 0 until arr.length()) {
-                    val seg = arr.optJSONObject(i) ?: continue
-                    val start = seg.optLong("start", -1)
-                    val end = seg.optLong("end", -1)
-                    if (end <= start) continue
-                    val durMin = (end - start) / 60.0
-                    when (seg.optString("stage")) {
-                        "wake"  -> aw += durMin
-                        "light" -> li += durMin
-                        "deep"  -> dp += durMin
-                        "rem"   -> rm += durMin
-                    }
-                }
-                if (aw + li + dp + rm > 0.0) StageMins(aw, li, dp, rm) else null
-            }
-            else -> null
-        }
-    }.getOrNull()
-}
-
-/**
- * Build the whole model from the cached daily metrics + the latest sleep session + the
- * export-verbatim sleep figures. Returns null when there is no usable latest night (no
- * stage minutes), which renders the empty state. All series are computed in one pass-set
- * here, matching the macOS buildModel(). Internal so SleepImportedFiguresTest can pin the
- * prefer-imported logic (the recoveryCalibrationNights test pattern).
- */
-internal fun buildSleepModel(
-    days: List<DailyMetric>,
-    session: SleepSession?,
-    imported: ImportedSleepSeries = ImportedSleepSeries(),
-    selectedDay: String? = null,
-    // The bridged main-night GROUP's summed stage minutes + full-night segments (#561), threaded from
-    // selectNight so a biphasic night's hero shows the WHOLE night, not one fragment (#555). Null for a
-    // single-block day → the session/DailyMetric path below is unchanged.
-    heroStages: StageMins? = null,
-    heroSegments: List<PersistedSegment>? = null,
-): SleepModel? {
-    val effectiveDay = selectedDay ?: days.lastOrNull()?.day ?: return null
-    // The HERO night = the selected day's stage-bearing row. The TILE / debt / need / trend
-    // window, by contrast, is the FULL history (latest-anchored) — matching iOS SleepView, which
-    // builds every tile series + the debt ledger + the personal need from `repo.days` regardless
-    // of which night the hero is browsing. Browsing a past night only re-points the hero, never
-    // the at-a-glance tiles or the "Last 14 nights" ledger. One cross-platform definition. (#5)
-    val latest = days.lastOrNull {
-        it.day == effectiveDay && (it.deepMin ?: 0.0) + (it.remMin ?: 0.0) + (it.lightMin ?: 0.0) > 0.0
-    }
-        ?: return null
-
-    // Prefer stage minutes from the session's (possibly reclipped) stagesJSON when it belongs
-    // to this night — so a wake-time edit on an imported or computed night updates stage cards
-    // (StagesVsTypical, Hypnogram footer) immediately without waiting on a rescore.
-    val sessionStageMins = session
-        ?.takeIf { AnalyticsEngine.dayString(it.endTs) == latest.day || localDayString(it.endTs) == latest.day }
-        // #259: trim to the EFFECTIVE onset before summing, so a hand-edited bedtime the raw was too sparse
-        // to re-stage (WHOOP 4.0) can't show pre-onset stages that push asleep past time-in-bed. No-op when
-        // the session already starts at its onset (the common case). Matches the analytics-side clamp.
-        ?.let { parseSessionStages(SleepStageTotals.clampStagesToOnset(it.stagesJSON, it.effectiveStartTs)) }
-    val deep = heroStages?.deep ?: sessionStageMins?.deep ?: latest.deepMin ?: 0.0
-    val rem = heroStages?.rem ?: sessionStageMins?.rem ?: latest.remMin ?: 0.0
-    val light = heroStages?.light ?: sessionStageMins?.light ?: latest.lightMin ?: 0.0
-
-    // Hero awake estimate works off ASLEEP minutes (totalSleepMin), never the in-bed window. The
-    // old code substituted the edited session's (wake − onset) duration — TIME IN BED — for the
-    // asleep figure here and across every per-tile pass, which inflated awake / hours-vs-needed /
-    // debt vs the actual sleep. iOS never did this (it derives awake straight from the decoded
-    // stage segments). Dropped for parity (#1/#7); a sleep edit now reaches the tiles via the
-    // re-score path, not a display-time in-bed swap.
-    val asleep = latest.totalSleepMin ?: (deep + rem + light)
-    // Awake estimate: prefer (time-in-bed − asleep) implied by efficiency; else from
-    // disturbances; matches the macOS "awake minutes" carried in the stagesJSON.
-    val effFrac = latest.efficiency?.let { if (it > 1.0) it / 100.0 else it }
-    val awake = when {
-        effFrac != null && effFrac in 0.01..0.999 -> max(0.0, asleep / effFrac - asleep)
-        latest.disturbances != null -> latest.disturbances * 6.0
-        else -> 0.0
-    }
-    val stages = Stages(awake = awake, light = light, deep = deep, rem = rem)
-    if (stages.total <= 0.0) return null
-
-    // Typical = mean across ALL nights with data (full history, latest-anchored — never bounded
-    // to the browsed night), mirroring iOS typicalTotalMin / typicalStageMin over repo.days.
-    val typicalTotalMin = mean(days.mapNotNull { it.totalSleepMin }.filter { it > 0.0 })
-    val typicalDeepMin = mean(days.mapNotNull { it.deepMin }.filter { it > 0.0 })
-    val typicalRemMin = mean(days.mapNotNull { it.remMin }.filter { it > 0.0 })
-    val typicalLightMin = mean(days.mapNotNull { it.lightMin }.filter { it > 0.0 })
-
-    // Personal sleep need (minutes): mean asleep, floored at 7.5h (450 min).
-    val needMin = max(450.0, typicalTotalMin ?: 450.0)
-
-    // Per-tile metrics — each a full pass over the FULL day history (asleep totals, no in-bed
-    // substitution), latest = the most-recent day. Mirrors iOS SleepView, where every tile series
-    // is `metric { … }` over repo.days. Where the WHOOP export carried the figure verbatim
-    // (metricSeries), it wins per day; the on-device recomputation fills the rest.
-    val performance = metric(days) { d ->
-        imported.performance[d.day]                       // WHOOP's own 0–100 figure wins per day
-            // else the REAL Rest composite (RestScorer.restFromDaily) — the SAME single source of
-            // truth the Today Rest score, the metric-detail overlay (below) and iOS SleepView
-            // (AnalyticsEngine.Rest.composite(daily:)) read. The old hours-vs-need proxy ceilinged
-            // live 5.0 nights at 100% here while every other surface showed the ~85% composite (#298).
-            ?: com.noop.analytics.RestScorer.restFromDaily(d)
-    }
-    val efficiency = metric(days) { d ->
-        d.efficiency?.let { if (it <= 1.0) it * 100.0 else it }
-    }
-    val consistency = run {
-        // Prefer the imported sleep_consistency series, but only when it covers the latest
-        // night — otherwise "latest" would silently be a months-old import-era value.
-        val lastDay = days.lastOrNull()?.day
-        if (lastDay != null && imported.consistency[lastDay] != null) {
-            val series = days.mapNotNull { imported.consistency[it.day] }
-            Metric(series.lastOrNull(), mean(series), series)
-        } else {
-            consistencySeries(days)
-        }
-    }
-    val hoursVsNeeded = metric(days) { d ->
-        val need = imported.needMin[d.day] ?: needMin   // imported need wins per day
-        d.totalSleepMin?.takeIf { it > 0.0 && need > 0.0 }?.let { it / need * 100.0 }
-    }
-    val restorative = metric(days) { d ->
-        val dp = d.deepMin; val rm = d.remMin; val sl = d.totalSleepMin
-        if (dp != null && rm != null && sl != null && sl > 0.0) (dp + rm) / sl * 100.0 else null
-    }
-    val respiratory = metric(days) { it.respRateBpm }
-    val sleepDebt = run {
-        val series = days.mapNotNull { d ->
-            imported.debtMin[d.day]   // minutes, export-verbatim
-                ?: d.totalSleepMin?.takeIf { it > 0.0 && needMin > 0.0 }
-                    ?.let { max(0.0, needMin - it) }   // APPROXIMATE fallback
-        }
-        Metric(series.lastOrNull(), mean(series), series)
-    }
-
-    // Trend set = the most-recent nights with data (asleep totals, full history — latest-anchored,
-    // not the browsed night). Mirrors iOS's trailing trend over repo.days.
-    val trendRows = days.filter { (it.totalSleepMin ?: 0.0) > 0.0 }.takeLast(14)
-    val trendHours = trendRows.mapNotNull { it.totalSleepMin?.let { minutes -> minutes / 60.0 } }
-    val trendNeedHours = trendRows.map { row -> ((imported.needMin[row.day] ?: needMin) / 60.0) }
-    val trendDebtHours = trendRows.map { row ->
-        val sleptMin = row.totalSleepMin ?: 0.0
-        val neededMin = imported.needMin[row.day] ?: needMin
-        ((imported.debtMin[row.day] ?: max(0.0, neededMin - sleptMin)) / 60.0)
-    }
-    val trendDates = trendRows.map { it.day }
-
-    // Real per-epoch timeline only when the merged session IS this night — UTC OR local-tz
-    // end-day match (imported DailyMetric.day is local-tz while dayString is UTC, so a
-    // near-midnight-UTC wake only matches via the local key; selectNight attributes the
-    // night the same way). A non-matching session degrades safely to synthesis, never to
-    // a wrong night. (#160)
-    val realSegments = heroSegments?.map { seg -> seg.stage to ((seg.end - seg.start) / 60f) }
-        ?: session
-            ?.takeIf {
-                AnalyticsEngine.dayString(it.endTs) == latest.day || localDayString(it.endTs) == latest.day
-            }
-            ?.let { parsePersistedSegments(it.stagesJSON) }
-            ?.map { seg -> seg.stage to ((seg.end - seg.start) / 60f) }
-
-    // Rolling 14-night sleep-debt ledger over the FULL day history (the analytics caps to the
-    // most-recent 14 counted nights and skips no-data nights), using the SAME personal need the
-    // tiles use (`needMin`, ≥ 7.5 h — the per-user override over the 8 h default). Full history,
-    // not the browsed-night window: the ledger is a "Last 14 nights" at-a-glance summary that
-    // matches the debt TILE (both now read asleep totals over `days`), and mirrors iOS's
-    // debtLedger over repo.days. (#242, #5)
-    val sleepDebtLedger = SleepDebt.ledger(
-        series = days.map { it.day to it.totalSleepMin },
-        needHours = needMin / 60.0,
-    )
-
-    return SleepModel(
-        stages = stages,
-        clockLabel = clockLabel(latest, session),
-        efficiencyText = efficiency.latest?.let { "${it.roundToInt()}%" } ?: "—",
-        performance = performance,
-        efficiency = efficiency,
-        consistency = consistency,
-        hoursVsNeeded = hoursVsNeeded,
-        restorative = restorative,
-        respiratory = respiratory,
-        sleepDebt = sleepDebt,
-        typicalTotalMin = typicalTotalMin,
-        typicalDeepMin = typicalDeepMin,
-        typicalRemMin = typicalRemMin,
-        typicalLightMin = typicalLightMin,
-        trendHours = trendHours,
-        trendNeedHours = trendNeedHours,
-        trendDebtHours = trendDebtHours,
-        trendDates = trendDates,
-        realSegments = realSegments,
-        sleepDebtLedger = sleepDebtLedger,
-    )
-}
-
-/**
- * #940 no-blank fallback: one impossible/stage-less SELECTED day (typically the newest, after a bad
- * hand-edit staged it all-awake) must not hide the whole tab's full-history surfaces. Re-anchor the
- * model to the newest day that HAS stage minutes; the tiles / ledger / trends it feeds are
- * full-history by construction, so this only changes which day supplies the hero-independent
- * anchor. Null only when NO day carries stage data (the true first-run empty state). Internal so
- * SleepPhantomNightFallbackTest can pin the rule. Mirrors iOS buildModel's stage-less stub fallback.
- */
-internal fun fallbackSleepModel(
-    days: List<DailyMetric>,
-    imported: ImportedSleepSeries = ImportedSleepSeries(),
-): SleepModel? {
-    val anchorDay = days.lastOrNull {
-        (it.deepMin ?: 0.0) + (it.remMin ?: 0.0) + (it.lightMin ?: 0.0) > 0.0
-    }?.day ?: return null
-    return buildSleepModel(days, null, imported, selectedDay = anchorDay)
-}
-
-/** Build a metric from a per-day transform, keeping only finite values. */
-private fun metric(days: List<DailyMetric>, transform: (DailyMetric) -> Double?): Metric {
-    val series = days.mapNotNull(transform).filter { it.isFinite() }
-    return Metric(series.lastOrNull(), mean(series), series)
-}
-
-/**
- * Consistency per day from the rolling bedtime spread — but Android's daily metrics carry
- * no per-night onset timestamp, so a bedtime-variance score isn't reconstructable from the
- * cached `days` alone. We approximate the same intent (steadier nights → higher score) from
- * the trailing-14 spread of total-sleep duration: low duration variability ≈ a consistent
- * routine. Each day's score uses the window ending at that day, matching the macOS rolling
- * shape. Honest note: this is a duration-based proxy, not the onset-spread score.
- */
-private fun consistencySeries(days: List<DailyMetric>): Metric {
-    val mins = days.mapNotNull { it.totalSleepMin?.takeIf { m -> m > 0.0 } }
-    if (mins.size < 3) return Metric(null, null, emptyList())
-    val scores = ArrayList<Double>()
-    for (i in mins.indices) {
-        val lo = max(0, i - 13)
-        val window = mins.subList(lo, i + 1)
-        if (window.size < 3) continue
-        val m = window.average()
-        val variance = window.sumOf { (it - m) * (it - m) } / window.size
-        val sd = Math.sqrt(variance)
-        // 90 min of duration SD maps to a 0 score; tighter routines climb to 100.
-        scores.add((100.0 * (1.0 - sd / 90.0)).coerceIn(0.0, 100.0))
-    }
-    return Metric(scores.lastOrNull(), mean(scores), scores)
-}
-
-private fun mean(vals: List<Double>): Double? = if (vals.isEmpty()) null else vals.sum() / vals.size
-
-// MARK: - Stage segment reconstruction (durations only — same architecture as macOS)
-
-/**
- * Lay the stage minutes end-to-end as proportional hypnogram segments: light → deep →
- * light → rem → light → awake (deep early, REM later, awake last). Weights are minutes;
- * the Hypnogram normalizes them to width.
- */
-private fun stageSegments(s: Stages): List<Pair<String, Float>> {
-    val out = ArrayList<Pair<String, Float>>()
-    fun add(name: String, minutes: Double) {
-        if (minutes > 0.0) out.add(name to minutes.toFloat())
-    }
-    add("light", s.light * 0.4)
-    add("deep", s.deep)
-    add("light", s.light * 0.3)
-    add("rem", s.rem)
-    add("light", s.light * 0.3)
-    add("awake", s.awake)
-    return out
-}
+// MARK: - Model + derivation lives in SleepModels.kt, SleepNightSelection.kt, SleepModelLogic.kt, and SleepStageTimelineLogic.kt
 
 // MARK: - Formatting helpers (mirror SleepView.swift)
 
@@ -3662,191 +2986,8 @@ private fun durationText(minutes: Double): String {
     return if (m < 60) "${m}m" else "${m / 60}h ${m % 60}m"
 }
 
-/** "Wed 4 Jun · 22:50–06:48" style trailing label from the session clock, when available. */
-private fun shortDayLabel(day: String): String =
-    runCatching {
-        LocalDate.parse(day).format(DateTimeFormatter.ofPattern("d MMM", Locale.US))
-    }.getOrDefault(day)
-
 private fun List<Double>.averageOrNull(): Double? =
     if (isEmpty()) null else sum() / size
-
-private fun clockLabel(latest: DailyMetric, session: SleepSession?): String {
-    if (session != null) return sessionClockLabel(session)
-    // Fall back to the daily metric's day string (YYYY-MM-DD), formatted to "EEE d MMM".
-    val dateFmt = SimpleDateFormat("EEE d MMM", Locale.US)
-    return runCatching {
-        val parser = SimpleDateFormat("yyyy-MM-dd", Locale.US).apply { timeZone = TimeZone.getTimeZone("UTC") }
-        parser.parse(latest.day)?.let { dateFmt.format(it) }
-    }.getOrNull() ?: latest.day
-}
-
-/** "Wed 4 Jun · 22:50–06:48" — the night-nav header's date · onset–wake line. (#160) */
-private fun sessionClockLabel(session: SleepSession): String =
-    clockLabelFor(session.effectiveStartTs, session.endTs) // EFFECTIVE onset so an edited bedtime shows (PR #395)
-
-/** Same date · onset–wake line from explicit unix-second bounds (the #736 group-aligned bedtime). */
-private fun clockLabelFor(onsetTs: Long, wakeTs: Long): String {
-    val timeFmt = SimpleDateFormat("HH:mm", Locale.US)
-    val dateFmt = SimpleDateFormat("EEE d MMM", Locale.US)
-    val onset = Date(onsetTs * 1000L)
-    val wake = Date(wakeTs * 1000L)
-    return "${dateFmt.format(onset)} · ${timeFmt.format(onset)} - ${timeFmt.format(wake)}"
-}
-
-/** Unix seconds → "YYYY-MM-DD" in the DEVICE timezone (vs AnalyticsEngine.dayString = UTC). */
-private fun localDayString(ts: Long): String =
-    SimpleDateFormat("yyyy-MM-dd", Locale.US).format(Date(ts * 1000L))
-
-/** Unix seconds → a local wall-clock "HH:mm" (same 24h formatting the nav-header span uses). */
-private fun clockTimeLabel(ts: Long): String =
-    SimpleDateFormat("HH:mm", Locale.US).format(Date(ts * 1000L))
-
-/** One persisted per-epoch stage segment (wall-clock unix seconds). */
-internal data class PersistedSegment(val start: Long, val end: Long, val stage: String)
-
-/**
- * Parse the verbatim per-epoch segments array the on-device stager persists
- * ([{"start","end","stage"}], unix seconds, stage ∈ wake|light|deep|rem — see
- * AnalyticsEngine.encodeStages). Returns null for the imported minutes shapes
- * (the macOS {"light",…} dict and the CSV-import [{stage,min}] array) and any
- * malformed input, so callers keep the synthesized fallback. Pure + unit-tested
- * (see SleepStageSegmentsTest).
- */
-internal fun parsePersistedSegments(json: String?): List<PersistedSegment>? {
-    if (json.isNullOrBlank()) return null
-    val trimmed = json.trim()
-    if (!trimmed.startsWith("[")) return null
-    return runCatching {
-        val arr = JSONArray(trimmed)
-        val out = ArrayList<PersistedSegment>(arr.length())
-        for (i in 0 until arr.length()) {
-            val o = arr.optJSONObject(i) ?: return@runCatching null
-            val start = o.optLong("start", Long.MIN_VALUE)
-            val end = o.optLong("end", Long.MIN_VALUE)
-            val stage = o.optString("stage", "")
-            if (start == Long.MIN_VALUE || end <= start || stage.isEmpty()) return@runCatching null
-            out.add(PersistedSegment(start, end, stage))
-        }
-        out.takeIf { it.size >= 2 }
-    }.getOrNull()
-}
-
-// MARK: - Stage timeline logic (iOS #988 port — pure, unit-tested)
-
-/** One contiguous run of a single sleep stage, in seconds from the night's onset. */
-internal data class StageInterval(val stage: String, val startSec: Double, val endSec: Double) {
-    val durationSec: Double get() = endSec - startSec
-}
-
-/**
- * Reconstruct absolute (stage, startSec, endSec) intervals from the hero's ordered
- * `realSegments` weight pairs (name, minutes) by walking cumulative fractions across [spanSec]
- * (design 2026-07-10, §Real-stage nights item 3). Non-finite / non-positive weights are skipped —
- * they carry no drawable width. Returns [] when nothing is drawable.
- */
-internal fun stageIntervalsFromWeights(
-    segments: List<Pair<String, Float>>,
-    spanSec: Double,
-): List<StageInterval> {
-    if (segments.isEmpty() || !spanSec.isFinite() || spanSec <= 0.0) return emptyList()
-    val weights = segments.map { (_, wt) -> if (wt.isFinite() && wt > 0f) wt.toDouble() else 0.0 }
-    val total = weights.sum()
-    if (total <= 0.0) return emptyList()
-    val out = ArrayList<StageInterval>(segments.size)
-    var cum = 0.0
-    segments.forEachIndexed { i, (name, _) ->
-        val w = weights[i]
-        if (w <= 0.0) return@forEachIndexed
-        val start = spanSec * (cum / total)
-        cum += w
-        out.add(StageInterval(name, start, spanSec * (cum / total)))
-    }
-    return out
-}
-
-/**
- * Display-time smoothing — a straight port of Swift `Hypnogram.displaySmoothed` (WHOOP-style,
- * Packages/StrandDesign/Sources/StrandDesign/Hypnogram.swift:92). The on-device stager emits
- * 30 s-epoch runs, so a real night arrives as 60–100 fragments; brief flickers are absorbed into
- * their surroundings AT DISPLAY TIME. Render-only: totals, percentages and stored data are
- * computed from the raw segments elsewhere and are untouched. Pass minDurationSec = 0 for raw.
- */
-internal fun displaySmoothed(
-    intervals: List<StageInterval>,
-    minDurationSec: Double,
-): List<StageInterval> {
-    // Match Swift `Hypnogram.displaySmoothed`: guard ONLY on count. A minDurationSec <= 0 ("raw") must
-    // still fall through to coalesce() — Swift returns the coalesced timeline, not the un-merged epoch
-    // fragments. Short-circuiting on <= 0 here returned 60-100 raw fragments (the "comb"), diverging from
-    // the port. With minDurationSec = 0 the absorb loop's `duration < 0` filter is empty, so it breaks
-    // right after the first coalesce — same result as Swift.
-    if (intervals.size <= 2) return intervals   // Swift: guard count > 2
-
-    // Coalesce adjacent same-stage runs (also bridges the zero-length seams between epochs).
-    fun coalesce(ivs: List<StageInterval>): MutableList<StageInterval> {
-        val out = mutableListOf<StageInterval>()
-        for (iv in ivs) {
-            val last = out.lastOrNull()
-            if (last != null && last.stage == iv.stage && iv.startSec - last.endSec < 1.0) {
-                out[out.size - 1] = StageInterval(last.stage, last.startSec, iv.endSec)
-            } else {
-                out.add(iv)
-            }
-        }
-        return out
-    }
-
-    var ivs = coalesce(intervals)
-    // Repeatedly absorb the shortest sub-threshold fragment into its longer neighbour,
-    // re-coalescing after each pass, until every remaining block clears the threshold.
-    while (ivs.size > 1) {
-        val idx = ivs.indices
-            .filter { ivs[it].durationSec < minDurationSec }
-            .minByOrNull { ivs[it].durationSec } ?: break
-        val victim = ivs[idx]
-        val prev = if (idx > 0) ivs[idx - 1] else null
-        val next = if (idx < ivs.size - 1) ivs[idx + 1] else null
-        when {
-            prev != null && next != null ->
-                // Absorb into the longer neighbour so the dominant surrounding stage wins.
-                if (prev.durationSec >= next.durationSec) {
-                    ivs[idx - 1] = StageInterval(prev.stage, prev.startSec, victim.endSec)
-                } else {
-                    ivs[idx + 1] = StageInterval(next.stage, victim.startSec, next.endSec)
-                }
-            prev != null -> ivs[idx - 1] = StageInterval(prev.stage, prev.startSec, victim.endSec)
-            next != null -> ivs[idx + 1] = StageInterval(next.stage, victim.startSec, next.endSec)
-            else -> break
-        }
-        ivs.removeAt(idx)
-        ivs = coalesce(ivs)
-    }
-    return ivs
-}
-
-/** Canonical stage key: trims, lowercases, and folds the "wake"/"awake" alias (stageColorFor parity). */
-internal fun canonicalStage(name: String): String {
-    val n = name.trim().lowercase()
-    return if (n == "wake") "awake" else n
-}
-
-/**
- * The (startFraction, widthFraction) spans of [rowStage]'s intervals within the night — one entry
- * per solid segment in that stage's timeline row track. Fractions of [spanSec]; the draw side
- * applies the min-width floor and canvas clamping.
- */
-internal fun stageRowSpans(
-    intervals: List<StageInterval>,
-    rowStage: String,
-    spanSec: Double,
-): List<Pair<Float, Float>> {
-    if (spanSec <= 0.0 || !spanSec.isFinite()) return emptyList()
-    val key = canonicalStage(rowStage)
-    return intervals
-        .filter { canonicalStage(it.stage) == key }
-        .map { iv -> (iv.startSec / spanSec).toFloat() to (iv.durationSec / spanSec).toFloat() }
-}
 
 // MARK: - Hours vs Needed card
