+++ android/app/src/main/java/com/noop/ui/SleepTimeLabels.kt
@@ -1,0 +1,47 @@
+package com.noop.ui
+
+import com.noop.data.DailyMetric
+import com.noop.data.SleepSession
+import java.text.SimpleDateFormat
+import java.time.LocalDate
+import java.time.format.DateTimeFormatter
+import java.util.Date
+import java.util.Locale
+import java.util.TimeZone
+
+/** "Wed 4 Jun · 22:50–06:48" style trailing label from the session clock, when available. */
+internal fun shortDayLabel(day: String): String =
+    runCatching {
+        LocalDate.parse(day).format(DateTimeFormatter.ofPattern("d MMM", Locale.US))
+    }.getOrDefault(day)
+
+internal fun clockLabel(latest: DailyMetric, session: SleepSession?): String {
+    if (session != null) return sessionClockLabel(session)
+    // Fall back to the daily metric's day string (YYYY-MM-DD), formatted to "EEE d MMM".
+    val dateFmt = SimpleDateFormat("EEE d MMM", Locale.US)
+    return runCatching {
+        val parser = SimpleDateFormat("yyyy-MM-dd", Locale.US).apply { timeZone = TimeZone.getTimeZone("UTC") }
+        parser.parse(latest.day)?.let { dateFmt.format(it) }
+    }.getOrNull() ?: latest.day
+}
+
+/** "Wed 4 Jun · 22:50–06:48" — the night-nav header's date · onset–wake line. (#160) */
+internal fun sessionClockLabel(session: SleepSession): String =
+    clockLabelFor(session.effectiveStartTs, session.endTs) // EFFECTIVE onset so an edited bedtime shows (PR #395)
+
+/** Same date · onset–wake line from explicit unix-second bounds (the #736 group-aligned bedtime). */
+internal fun clockLabelFor(onsetTs: Long, wakeTs: Long): String {
+    val timeFmt = SimpleDateFormat("HH:mm", Locale.US)
+    val dateFmt = SimpleDateFormat("EEE d MMM", Locale.US)
+    val onset = Date(onsetTs * 1000L)
+    val wake = Date(wakeTs * 1000L)
+    return "${dateFmt.format(onset)} · ${timeFmt.format(onset)} - ${timeFmt.format(wake)}"
+}
+
+/** Unix seconds → "YYYY-MM-DD" in the DEVICE timezone (vs AnalyticsEngine.dayString = UTC). */
+internal fun localDayString(ts: Long): String =
+    SimpleDateFormat("yyyy-MM-dd", Locale.US).format(Date(ts * 1000L))
+
+/** Unix seconds → a local wall-clock "HH:mm" (same 24h formatting the nav-header span uses). */
+internal fun clockTimeLabel(ts: Long): String =
+    SimpleDateFormat("HH:mm", Locale.US).format(Date(ts * 1000L))
